Run Importer updates
Fix run control file "stop run" to actually wait for the import to
finish.
Add start script that writes PID to a file.
Add script that checks process with PID in same file and restarts the
process if not running.
Update pgm to remove the PID file when exit using "stop run"