Single Prot:Trigger Update Width and Position using MutationObserver
Trigger Update Width of main DIV and Position of Boxes on upper right
using MutationObserver.
The width of the main div ( fake 'overlay') and position of Boxes on
upper right has complicated rules so use Javascript for width and
positioning.
Updates to the Reported Peptide List can change the width which then may
change the positioning so put MutationObserver on Reported Peptide List.