Fix Bug that is NOT currently causing invalid data to display
The data for this is being correctly displayed, due to how the code runs
for all searches before all the AJAX responses.
In ProteinViewPage_DisplayData_SingleProtein_SingleSearch_LoadProcessDataFromServer.loadDataForInitialOverlayShow(...)
If the protein sequence is already loaded, then
this._populateStaticModificationsPositionsOnProteinSequence(...) is not
called.
Changed to call
this._populateStaticModificationsPositionsOnProteinSequence(...)
in that case.