Convert to Typescript - Many converted, Many still Javascript
Add Typescript to the limelight_webapp/front_end

Convert Many Javascript files to Typescript.
* The conversions were minimal to get the files to compile under
Typescript along with typing some parameters and variables that are of
classes defined in Limelight webapp.

Many Javascript files were not converted to Typescript.

Some work-arounds were used to get Javascript to compile.
All uses of 'require' (CommonJS and/or AMD Imports) were moved to
separate Javascript files and the require results were then exported and
imported into the Typescript code.  This was done since most code uses
ES6 'import' and so that was the chosen module system for Typescript.