Importer: Allow Scan Filename on PSM in input to not have Suffix
The Importer when matching the Scan Filename on PSM in input Limelight
XML file to the scan filenames being imported will:
first attempt to match the Scan Filename on PSM to scan filenames being
imported.
If not found, the Importer will attempt to match the Scan Filename on
PSM to scan filenames being imported with the filename suffix removed.
ie: 'aa' on PSM will match aa.mzML being imported.

The webapp and the submitter program will not allow duplicate scan
filenames to be submitted and will extend that to duplicate scan
filenames when the suffix is removed.