Web: Add Spinner React Component, Add it to Single Protein Loading
Add new <Spinner_Limelight_Component/> that adds the spinner on mount
and removes it on unmount to <div> in the component.
Position the component to where the spinner should be.

  new functions were added to spinner.js that accept a HTMLDivElement
DOM node
   A new class SpinnerManagerClass_Limelight was added to spinner.js
that tracks it's own instance of the result of calling spinner.spin()
for calling stop() on.
This Component uses SpinnerManagerClass_Limelight
so it operates independently of calls to existing functions in
spinner.js (createSpinner, destroySpinner, incrementSpinner,
decrementSpinner)