Runnable Jars change for Java 9+ for log4j2
For Jars: Importer, Run Importer, Submitter
Gradle shadowJar is used to create a runnable jar
so the manifest file of the included exploded jars are not used.
The create manifest file for each jar now includes
    "Multi-Release": true,
So that log4j2 does not cause the following warning to be displayed
WARNING: sun.reflect.Reflection.getCallerClass is not supported. This
will impact performance.