Single Protein: Back button restores URL and reload to show Protein List
This is for Protein Page: Single Search, Single Search Sub Groups,
Multiple Search, Experiment.
Clicking the back button is not the same as clicking the "X" to close
Single Protein.
User selection changes in Single Protein can alter user selection
changes on the Protein list, and when click "X" to close Single Protein
those are preserved on the Protein List.
But when click Browser Back Button the URL that was current immediately
before the single protein was clicked is pushed to Browser History and
when Browser Back button is clicked that URL is restored as current and
the page is reloaded.