Peptide/Exp-Peptide: Fix Show Open Mod mass On Page Load
Peptide Page and Experiment Peptide Page
On Page Load when "Show in Peptides:" "Open Modifications" checkbox is
selected (which is default) the Open Modification PSM masses were not
loaded
and so the Peptides in the Peptide List did NOT contain the Open Mod
Mass.
This has been fixed.

Also fixed is Experiment Peptide page not loading most Page state on
page load.
Also fixed is Experiment Peptide page error on select "Treat Mass 0 As
Unmodified:"