ModPage: [Download ZScore Report],etc: nest in <span> to limit clickable
Before each entry was in <div> which made clicking to the right of each
entry also click that entry.
Now each entry is nested in a <span> inside it's <div> and that <span>
is what is clickable so now only clicking on the text of each entry will
run the click event.