Project: Share Data: Add Reviewer Mode and update the block on page
In the Project Authentication, add a Reviewer Mode.
When Reviewer Mode (AKA Public Access Code) set on a project and
enabled, accessing the project when not logged in will change the signin
page to request a Reviewer Access Code instead, with a link to switch to
normal Limelight Signin.

Share Data block on Project Page:

Add:
Enable Reviewer Mode and generate new Reviewer Access code

Update overall layout and functioning of the Share Data block.