Add Project Search Id Code.  For opening data pages from Project page
Add a Project Search Id Code string of capital letters that is unique
for each Project Search Id.

Before, when open peptide, protein, mod page from Project page,
a new tab was opened empty which could left empty if the webservice call
to get the searchDataLookupParamsCode failed or forced the project page
to reload since the user session had expired.
Now, open peptide, protein, mod page in new tab directly using the
Project Search Id Code string and on the server the
searchDataLookupParamsCode will be created and passed in the HTML and
the browser URL will be updated by the JS to have the
searchDataLookupParamsCode.