Web: Upload Files for Import: Split into chunks
For the web app only (for now).
Split files sent from the browser via JS to the server into chunks
Send 1 chunk at a time to the server and wait for completion before send
next chunk.
Retry send a chunk a few times when get network failure error.
Immediately fail on HTTP status code other than 200.
On server side, the newly arrived chunk is appended to the end of the
file, unless the file is already long enough to include the chunk.