Reviewer Access Code: show input when project not for current code
Reviewer/Public Access Code
When User Session is for Public Access Code and user tries to access a
project that is not for the Project that the Reviewer/Public Access Code
in the session is for,
Show the signin page which will show the input field for the Public
access code.
When the user enter the Reviewer/Public Access Code for that project,
then the user can access that project
BUT loses access to the project the previous Reviewer/Public Access Code
in the User Session was for.