Project Short Name: Set null when mark for deletion
When a user deletes a project, it is not deleted.
A flag on the project table record is set to 1(true) that the record is
to be deleted.
When the user deletes the project, the expectation is that the short
name is available again so the short name on the updated project record
is set to null.