Project Public - Access issues: logged in & Reviewer Access Code Session
Access control was More restrictive so no data was shared that was not
supposed to be shared.

When a project is public:
User that is logged in with account that does not have access to that
project is not able to view that project even though it is public.
User that used a Reviewer Access code for one project is not able to
view another project that is public.