"Show in Peptides:" Single Protein: use default selection on open
The "Show in Peptides:" for Single Protein will always use default
selection when open the Single Protein.
The current selection for "Show in Peptides:" will be persisted to the
URL so any reload or load of the URL of Single Protein will retain the
selection chosen.
Closing the Single Protein and then open Single Protein will result in
default selection for "Show in Peptides:", even for same protein.