Web: Add "Viewer" Read only account type for Project Access Type
When adding an account to a project, there is a new type of "viewer"
which will have only Read access to that account.
Accounts of that type will not be able to make any additions or changes.
This type is more restrictive than type "Researcher".
Type "Researcher" which can make some changes and additions to changes
to the project.