Web: Reported Peptide & PSM Lists: Use SearchDataLook... Ann Display Ids
Existing Displayed data is correct since currently no way to change
Annotation Type Ids displayed so the existing displayed default data is
correct, for now.
Fix so For Tables that list Reported Peptide & PSM data that the
displayed Annotation Types are driven by the Annotation Type Ids to
Display listed in the SearchDataLookupParameters_Root are used.