User can change displayed Annotation Types (Reported Peptide and PSM)
For Reported Peptide and PSM lists, users can now change which
Annotation Types values are displayed.
The user can click "Change Displayed Peptide and PSM data" and make the
changes in the overlay.
The changed values are stored in class SearchDataLookupParameters_Root.
The user can change these values on the Peptide and Protein pages and on
the Single Protein overlay.
The PSM table list on the Mod page is impacted by the user selections.