QC Page rendering error in Firefox: Remove "inline-block" from <div>
The QC page has a rendering problem in Firefox where Firefox appears to
just stop rendering page when it gets to the <div> that is the container
of the chart sections.
That <div> had: display="inline-block"
The display="inline-block" was for minor formatting.
The display="inline-block" was removed so that the page would render in
Firefox.