Add tooltips to recently added "Filter On ..." entries
Add tooltips to "Filter On ...":

Filter On Scan Filename: Only data from the selected scan file(s) will
be used.

Filter on Retention Time (Minutes): Only PSMs with a retention time in
the supplied range will be used.

Filter on Precursor M/Z: Only PSMs with a MS1 precursor ion with an
observed m/z in the supplied range will be used.