webapp: bugfix: NO Invalid data displayed to user: threw a JS Error
Other code that checked for PSM Id Set size zero threw an Error so no
invalid data displayed to the user.
Bug is when intersection of PSM Ids for a Reported Peptide Id results in
no PSM Ids.
Now that Reported Peptide Id will NOT be in result of get Reported
Peptide Ids and their PSM Ids for filters.