Web: Bugfix: Related to Spring MVC/Boot version update
Default for returning text/message in exceptions with annotation
@ResponseStatus(
  value = org.springframework.http.HttpStatus.BAD_REQUEST,
  reason = "TEXT_HERE"
)
was changed to NOT send the reason text.

The Javascript in the Limelight application uses the reason text to
determine how to handle the returned non 200 status code so the reason
text must be returned.

Updated file application.properties adding:
server.error.include-message=always