Submitted Import files removal: Remove after delay when configured
When Delete successful import is selected in configuration, the failed
imports were never deleted.
Now they will be deleted after 3 days.
Add second selection for Successful and Failed to be deleted after 3
days.
Also when either is selected, any import in Started status for 15 days
the files will be deleted.
This all runs in the Run Importer process.