Web:ParallelStream Code Fix and improvements
Changed code to be compatible with ParallelStream:
Changed Java class SearcherCutoffValuesSearchLevel to not change
anything on 'get...' call.  Update the List when the Add... is called.

In all uses of ParallelStream, for all variables used from outside the
stream put them in an object with 'volatile' on all properties then use
that object inside the stream.