Webapp: Single Protein Experiment: Protein Coverage for Filters
Update code that triggers the RED colored sequence positions to identify
the sequence covered by the filtered data
to be triggered by ALL the filter State Objects.

Add a reminder to update that in
SingleProtein_FiltersDisplay_ComponentData
which is also updated when add a new filter.

In Single Protein: Protein Coverage for Filters:
Check that scan_RetentionTime_MZ_UserSelections_StateObject has a value