Submit Import:Add:'--misc-file=' but it is currently commented out
Commented out until '--misc-file=' is approved.

Accept new 'miscellaneous' files that are stored and then can be
downloaded from the search under search details 'Additional files:'.

These files are stored in the "File Object Storage Service" and are only
accepted for import when the "File Object Storage Service" is configured
in Limelight.