Bugfix:Web: Protein Position Filter cleanup for missing protein ids
The cleanup already existed but was not working properly.
The State object was not updated correctly.
The updated State object was not persisted to the URL.

class ProteinPositionFilter_UserSelections_StateObject

For the Protein Position Filter, remove Protein Sequence Version Id
entries that are not in the the current Searches/Search Filters
(PSM,...)
This applies to pages: Peptide, Peptide Experiment, QC.

Protein Sequence Version Ids may not be in the search results due to
changes of searches or changes in Search Filters (PSM,...).