Web:Mod Page:Remove user drag search to change order in graphic
Remove the ability to change the search order by dragging the searches
in the graphic.
Updated search order stored in the URL for existing URLs will be
retained (which is all URLs once user clicks "Update Visualization" or
drags to change search order).
For all new page views, the search order is never stored in the URL page
encoding.