Bugfix:Web: Get ReportedPeptideIds: NO PSM Filters & PSM has Decoy flag
Bug in the common server side code to get Reported Peptide Ids for
PSM/Reported Peptide/... filters.

PROBLEM:

For When the filters NOT match the default filters
AND there are NO PSM Filters
AND the search contains PSMs with the Decoy flag:
the get Reported Peptide Ids will return Reported Peptide Ids where the
ONLY PSMs have the Decoy flag.

Changed so that if the Search has flag that it contains PSMs with Decoy
flag that the get Reported Peptide Ids when have NO PSM Filters it will
get the PSM count for each Reported Peptide Id.