Bugfix: Web:QC:Digestion:
Fix:
For search with Open Mods and “Distinct Peptide Includes:“ does NOT have
“Open Modifications“ checked, still process the Open modifications for
determining which Missed cleavages to not count.

Also changed:

Missed cleavages are not counted if there is a modification at that
position.
For Open modifications unlocalized (without a position) it was coded to
not count any missing cleavages.
Change to ignore Open modifications unlocalized (without a position) .