Run Importer: Delete uploads that were never submitted
Delete when the upload was started but never the final 'submit' was sent
to mark the upload as complete and submitted.
Do this delete after 15 days after the upload started.

Delete is for all upload types in both tables:
file_import_tracking_tbl
import_and_pipeline_run_tracking_tbl

Check is for status INIT_INSERT_PRE_QUEUED.
When submit is called, status changes to QUEUED.

This delete after 15 days is the same as the current one for status
STARTED which is after start of the processing of the upload.