Bugfix: Webapp:Single Protein:Filter on Peptide Sequence
Filter on Peptide Sequence was executing different code for Single
Protein that did not filter in the same way as on the main page and not
consistent with the help text.

Changed Filter on Peptide Sequence for Single Protein to use same code
as main page.  The input reported peptides are already filtered for the
single protein so this is correct.