Fix commit 2 prev: Webservice URI version num incr for chng retrn data
In commit 2 previous:

A Webservice was changed to return data in different format.
This is webservice return data is cached so it has a version number at
the end of the URI.
The Java property holding the URI was renamed but the URI itself was not
updated.
Now the URI is updated along with the JS code calling it.