BUGFIX: PSM Download Contents: fix when search NO Scan Data
PSM Download Contents when click "Download All PSMs" on Peptide and
other pages.

When the search does NOT contain scan data (No Scan files imported with
the search)
the outputting of the section for field "SPECTRUM VIEWER URLS (comma
delim)" for each data line was outputting 2 tab characters '\t' instead
of 1.
This resulted in all columns following in the data rows to be shifted 1
column to the right.

This has now been fixed.