Mod Page: Change Object used as Map to actual Map:
Typscript 'class ModView_VizOptionsData_SubPart_selectedStateObject'

The property 'data' was an Object that was used as a Map with the keys
being the ProjectSearchIds.

Rename property 'data' to 'data__ModMass_Array_Map_Key_ProjectSearchId'
and change the type from 'any' to Map<number, Array<number>>
The key is ProjectSearchId and the value is array of selected Mod
masses.

Update the encode/decode to/from page state URL to create the new
structure for 'data__ModMass_Array_Map_Key_ProjectSearchId'