Mod Page: Add try/catch around 'async' functions to report exception
Do to report the exception in the function it occurs to the server.
async functions turn the exception into Promise reject which is turned
into exception when await.

Changed report exception to user and to server to ignore exception of
Webservice Error Class.

Added some typescript
Renamed some private methods/functions to start with '_'

Removed testing code