Web: Mod View Page: Rewrite to use filters and loaders from Peptide Page
Rewrite the Modification View Page.

Add the filters and options on the Peptide page under
"Click to Show Filters and Options".

Rewrite the rest of the Modification View Page to use the output from
the filtering under "Click to Show Filters and Options"
and use the common load data from webservices that the
Peptide and Protein pages use.