Bugfix: Mod Page: Fix Color in Viz when use "Max cutoff for color scale"
On the Mod Page
When enter value for "Max cutoff for color scale" for Count or Ratio,
limit the "Min" and "Max" passed to colorScale function returned from
'd3' to the min and max values passed to d3 to make that colorScale
function.

Previously, numbers outside the min/max were passed resulting in colors
like black.