Web: Protein Position Filter: NO Webservice call when NOT filtering
The Protein Position filter was making Protein data webservice call
when it was mounted on the main page
when NO proteins were being filtered on.

Protein Position Filter used in
"Filter On Protein Position:"
"Filter on Modifications at Protein Position:"

Code changed to ONLY make Protein data webservice call
when mount on main page when filtering on proteins and thus
need the Protein data.
If the protein data is not already loaded, when click "Add" and open the
overlay "Select Proteins" the Protein data is loaded at that time.