Web:BUG Fix: PSM Download: Fix download for Experiment
PSM Download for experiment was failing for Java null pointer exception.
The property 'conditionGroupLabel' was not passed from the Javascript so
in the Java when attempting to write that property which was null to the
output resulted in Java null pointer exception.

Changed the Javascript code to pass the property 'conditionGroupLabel'.

Added Java code to log and throw exception if property
'conditionGroupLabel' is null.

Added Java code so that if there an exception while writing the result
that it will attempt to write this message at the end:

"ERROR creating file.  Please contact the Limelight administrator."