Web: Protein Grouping: Change to have Promise about every 3 seconds
The promise will probably allow reading the UI about every 3 seconds in
future coding.
May need to add setTimeout.

This is a first step to maybe changing the code to allow a prompt the
user to change to Sub Groups option "Simple" or "No Groups" and being
able to read that input.