Web: Protein Grouping: Optimization when Every peptide in ONLY 1 Protein
When Every distinct peptide is in only 1 protein, then every protein is
distinct and will not be a subset of another protein
(Using either option of 'Parsimonious' or 'No Subgroups')
so put each protein in it's own protein group.

This helps with Casanovo data and may help with other similar data
that does not have proteins so create a fake protein for each peptide.