Support Searches with NO Proteins: XSD/DB/Importer/Web App
Some searches like Casanovo do not have proteins.
The Casanovo converter creates proteins.

Change Limelight to no longer require proteins.

The Limelight import file XSD has been changed.

The DB is changed.
The Importer is changed.

The webapp is changed:
If the search does not have proteins:
  The protein page cannot be displayed.
  If an existing URL for protein page is used a message is displayed.
  No Protein data is displayed on the other pages.

All of this applies if any of the searches in a "Compare" or an
Experiment do NOT have proteins.