Web: Data Table: Show at end of table: 'Currently Showing:'
For the Limelight home grown <<DataTable_TableRoot>

When a table has paging and more than one page,
at the start of the table is 2 lines:
'Currently Showing:' ...
and the next one.

NOW
display those 2 lines at the end of the table as well.
This is to inform the user that there are other pages of data for that
table.