Build/Web: Update to get release from Env Var, GH Actns set from release
Start of displaying the release in the web app.

Update the Github action to build to set environment variable
'LIMELIGHT_RELEASE_TAG' from the GIT release tag.
Update the web app build to get the release string from an environment
variable 'LIMELIGHT_RELEASE_TAG'
and store it in the file with the GIT tag and hash.
Update the web app on startup to read that file and store the release
string.
Update the web app on all pages to the HTML from the JSP to include a
<script> tag with the release string.
Will later find a place to display the version to the user.