Web: PSM Tbl Webservice: remove force output arrays with no data
Testing code cleanup.
Remove testing code that force output arrays with no data.

Data returned from the webservice was correct so this is not a bugfix.

Data returned from the webservice just had data that if wasn't in the
output would have the default values applied in the JS code would be the
same as what was in the webservice output.

The JS code already handles when the optional arrays are not populated.