Web:Mod Page: Residues: Show graphic of Mod count per residue letter
In many places where the residue letters were listed
replace with a SVG graphic showing modification count per residue
letter.

Also changes in that unlocalized modifications are applied to the
residue letters as a fraction of '1' for each peptide position.

When the filter "Filter on Modifications at Protein Position:" is used,
the modification positions or positions of the peptide that pass the
filter are the only ones used
but the divisor is the total count of positions or total length of the
peptide.