Web:Front end Error logging: Not log parsing URL when contain '/go/'
It appears that some crawlers do not allow
window.history.replaceState
to update the URL in the browser.

When change from URL shortcut with '/go/' then replace the URL in
browser with actual URL
the Front end code assumes it can read the URL in the browser and parse
it like normal but finds the URL shortcut with '/go/' instead.

No reason to log this to the server so check for that specific error
with new exception class and skip log to server.