Web: Submit Import File Upload: web upload only add data Checksum check
Only for the Submit Import File Upload via the web app.

At this time not updated for
Submit Import File Upload via Submit Import program.

Compute SHA-256 sum and pass that with each file chunk.
Compute same on server and compare the SHA-256 sum.
Reject the file chunk if not match.

The client Javascript will retry the send of the file chunk up to 4
times
(same counter as already used for retry on HTTP connection fail).