Web:QC Page: Centralize setting 'chart_Layout.hovermode = false'
Setting 'chart_Layout.hovermode = false' in Plotly Layout to prevent
mouse hover causing 100% CPU for certain chart type of ScatterGL when
high number of points on top of each other.

This is a fix for a known Plotly problem.

A potential fix is this
https://github.com/plotly/plotly.js/issues/5927#issuecomment-1697679087

but it is unclear if that will fix for all types of data.
For now adding
'spikedistance: 0 '
to the layout object but still disabling hovermode.