Web: Chromatogram Options: Change Retention Time to use overlay
Change so user enters modified retention time values in an mini overlay.

A small overlay positioned on top of the existing retention time values.

This brings it in line with rest of Limelight where if need to click a
button for values to be used that it is in an overlay.

It had the inputs on the main page and the user had to remember to click
the "Update" button for the changes to take effect.